map-function/requirements.txt
# Function dependencies, for example:
# package>=version
nltk
requests
google-cloud-storage